[#55] refactor : PR review 수정
