+++ src/components/TextField/TextArea/TextArea.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, useState } from 'react'
+import React, { useState } from 'react'
 import { TextAreaContainer, TextAreaWrapper, Textarea } from './styles'
 import { ReactComponent as Delete } from '../../../assets/delete_textfield_24.svg'
 import { ErrorText } from '../DefaultTextfield/styles'
@@ -12,12 +12,10 @@ interface TextAreaProps {
 
 const TextArea = ({ value, setValue, placeholder, error, errorMsg }: TextAreaProps) => {
   const [isFocused, setIsFocused] = useState(false)
-  const handleInputChange = useMemo(
-    () => (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-      setValue(event.target.value)
-    },
-    [setValue, value],
-  )
+  const handleInputChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
+    setValue(event.target.value)
+    console.log('value:', value)
+  }
   const onDelete = () => {
     setValue('')
   }
